QStabilizerHybrid (#484)
* Scratch work, stabilizer-inspired QUnit-type algorithm

* More QUnit-/stabilizer-inspired CNOT cases

* More stabilizer-inspired cases

* CNOT X control cases

* 'Fold-over' CNOT case

* Other CNOT cases

* Debugging CNOT

* Systematically debugging stabilizer-inspired CNOT

* Random Y sign

* Debugging identity 'local phase'

* Expanding QStabilizer API

* Typing; Append/Remove

* Y statistics

* #367: Handle |++>/|--> with Y generator

* Fix PAULI_MASK

* Adapting CHP as QStabilizer port for C++/Qrack

* Debugging and removing junk

* Debugging

* Fix segfault

* Add missing return statement in ApplySingleInvert

* Debugging setBasisState()

* Fix controlled edge cases

* Debugging Decompose

* Debugging compose

* Debugging Compose()

* Fixing Compose() segfault

* Debugging Compose

* Debugging Compose/Decompose segfaults

* Debug ProbAll/ProbMask

* Fixed Compose()

* Fixing test_swap_shunts

* Fixed Decompose()

* QUnit integration

* Remove isClifford (will be redundant)

* Fix EndEmulation for stabilizer (todo: other bases)

* Debugging QStabilizerHybrid

* Async dispatch for QStabilizerHybrid

* Respecting QUnit CPU parallel CMake option

* Debugging QUnit::Decompose()

* QStabilizer refactor

* Refactoring Compose/Decompose

* Refactor QStabilizer

* Reuse vectors in SetPermutation()

* Reuse sub-vectors

* Refactor

* No need to init r vector

* std::fill rather than assign

* Refactor

* Revert isClifford removal

* Debugging QUnit pointers

* Fix Compose() qubitCount

* Edge cases

* Fix uninitialized conditionals

* QINTERFACE_RESTRICTED in unit tests

* Fix typo

* Compose/Decompose safety

* QStabilizer::GetQuantumState() base case

* test_base_case

* Debugging QStabilizer test_base_case

* Debugging base case

* Temporarily disable QUnit -> QStabilizerHybrid unit tests

* Debugging test/benchmark

* gaussian() before Decompose()

* test_iswap is passing

* Hack to prevent unnecessary stabilizer init

* Fixing uniformly-controlled; optimizing TrySeparate()

* Debugging arithmetic isClifford

* Debugging unit tests

* Fix off-by-one in Compose()

* Debugging unit tests

* Revert QUnit

* Revert "Revert QUnit"

This reverts commit 1a0ceb59e3fd10270790709a3dba2b475e5aa7f1.

* Decide isClifford before EndEmulation()

* Remove isClifford

* Reverting SetQuantumState bool hack

* Debugging ForceM()

* Fixed ForceM()

* Simplify M() random (test still broken)

* Remove unnecessary gaussian() call

* Fix IsSeparableY() bug

* Bug is in ForceM

* Refactor ForceM()/M()

* Debugging QStabilizerHybrid() optimized gates

* Debugging QStabilizer::M()

* Debugging M()

* test_forcem is passing

* Debugging test_forcem

* Debugging CNOT

* Switch QUnit/QStabilizerHybrid layer order

* Optimizing

+++ test/tests.hpp
@@ -75,13 +75,13 @@ inline std::ostream& outputProbableResult(std::ostream& os, Qrack::QInterfacePtr
 {
     bitCapInt i;
 
-    double maxProb = 0;
+    float maxProb = 0;
     bitCapInt maxProbIdx = 0;
-    double totalProb = 0;
+    float totalProb = 0;
 
     // Iterate through all possible values of the bit array
     for (i = 0; i < qftReg->GetMaxQPower(); i++) {
-        double prob = qftReg->ProbAll(i);
+        float prob = (float)qftReg->ProbAll(i);
         totalProb += prob;
         if (prob > maxProb) {
             maxProb = prob;
